feat: integrate imageProcessEngine with web application
The imageProcessEngine uses yolo image detection, simpleHTR text detection, and a simple mapping algorithm to match textboxes with their labels. The web application provides a user interface to upload an image which will be run through our imageProcessEngine. The middle step is that the results of imageProcessEngine is run through the generateApi.py code to create the final web application that is based on the user submitted photo.